Add files via upload
This is the first draft of the code. 

workflow1()  - tests the gRPC endpoints, but the Say..() methods do not do what is expected. I would appreciate help on what is incorrect.
workflow2() - reads from the vault and un-marshals the results
workflow3() - tests List() and Read() methods on the Vault, also creates a certificate and reads it for verification.